+++ src/components/util/util.test.js
@@ -18,7 +18,21 @@ import {
   retrieveNoteReferredEntityDataFromLocationState,
   getClosedRequestStatusesFilterString,
   getOpenRequestStatusesFilterString,
+  getCentralTenantId,
+  isConsortiumEnabled,
 } from './util';
+
+const STRIPES = {
+  hasPerm: jest.fn().mockReturnValue(true),
+  hasInterface: jest.fn().mockReturnValue(true),
+  user: {
+    user: {
+      consortium: {
+        centralTenantId: 'test'
+      }
+    }
+  }
+};
 
 describe('accountsMatchStatus', () => {
   it('returns true if all accounts match', () => {
@@ -362,5 +376,29 @@ describe('getContributors', () => {
   it('getOpenRequestStatusesFilterString  ', () => {
     const data = getOpenRequestStatusesFilterString();
     expect(data).toStrictEqual('requestStatus.Open - Awaiting pickup,requestStatus.Open - Awaiting delivery,requestStatus.Open - In transit,requestStatus.Open - Not yet filled');
+  });
+});
+
+describe('isConsortiumEnabled', () => {
+  it('should return empty string', () => {
+    const data = isConsortiumEnabled();
+    expect(data).toBeFalsy();
+  });
+
+  it('should return centralTenantId', () => {
+    const data = isConsortiumEnabled(STRIPES);
+    expect(data).toBe(true);
+  });
+});
+
+describe('getCentralTenantId', () => {
+  it('should return empty string', () => {
+    const data = getCentralTenantId({});
+    expect(data).toBe('');
+  });
+
+  it('should return centralTenantId', () => {
+    const data = getCentralTenantId(STRIPES);
+    expect(data).toBe('test');
   });
 });