+++ src/views/UserSearch/Filters.js
@@ -4,24 +4,22 @@ import {
   FormattedMessage,
   injectIntl,
 } from 'react-intl';
-import {
-  get,
-} from 'lodash';
+import { get } from 'lodash';
 
 import {
   Accordion,
   AccordionSet,
   FilterAccordionHeader,
 } from '@folio/stripes/components';
-
+import { stripesConnect } from '@folio/stripes/core';
 import {
   CheckboxFilter,
   MultiSelectionFilter,
 } from '@folio/stripes/smart-components';
 
-import { statusFilter } from '../../constants';
-
 import CustomFieldsFilters from '../../components/CustomFieldsFilters';
+import { isConsortiumEnabled } from '../../components/util';
+import { USER_TYPES, statusFilter } from '../../constants';
 
 class Filters extends React.Component {
   static propTypes = {
@@ -34,6 +32,7 @@ class Filters extends React.Component {
     resultOffset: PropTypes.shape({
       replace: PropTypes.func.isRequired,
     }),
+    stripes: PropTypes.object.isRequired,
   };
 
   static defaultProps = {
@@ -85,15 +84,38 @@ class Filters extends React.Component {
       onChangeHandlers: { clearGroup },
       intl: { formatMessage },
       resources,
+      stripes,
     } = this.props;
     const {
       active = [],
       pg = [],
       tags = [],
       departments = [],
+      userType,
     } = activeFilters;
 
     const departmentsAreNotEmpty = !!resources.departments?.records?.length;
+
+    const isConsortium = isConsortiumEnabled(stripes);
+    const { PATRON, SHADOW, STAFF, SYSTEM } = USER_TYPES;
+    const userTypeOptions = [
+      {
+        value: PATRON,
+        label: formatMessage({ id: 'ui-users.information.userType.patron' }),
+      },
+      {
+        value: STAFF,
+        label: formatMessage({ id: 'ui-users.information.userType.staff' }),
+      },
+      {
+        value: SHADOW,
+        label: formatMessage({ id: 'ui-users.information.userType.shadow' }),
+      },
+      {
+        value: SYSTEM,
+        label: formatMessage({ id: 'ui-users.information.userType.system' }),
+      }
+    ];
 
     return (
       <AccordionSet>
@@ -161,6 +183,25 @@ class Filters extends React.Component {
             aria-labelledby="users-filter-accordion-tags"
           />
         </Accordion>
+        {
+          isConsortium && (
+            <Accordion
+              displayClearButton
+              id="users-filter-accordion-userTypes"
+              header={FilterAccordionHeader}
+              label={formatMessage({ id: 'ui-users.userType' })}
+              separator={false}
+              onClearFilter={() => clearGroup('userType')}
+            >
+              <CheckboxFilter
+                dataOptions={userTypeOptions}
+                name="userType"
+                selectedValues={userType}
+                onChange={this.handleFilterChange}
+              />
+            </Accordion>
+          )
+        }
 
         <CustomFieldsFilters
           activeFilters={activeFilters}
@@ -172,4 +213,4 @@ class Filters extends React.Component {
   }
 }
 
-export default injectIntl(Filters);
+export default stripesConnect(injectIntl(Filters));